+++ src/RandomInitialization/XavierInitialization.cpp
@@ -24,8 +24,8 @@ void XavierInitialization::initialize(Eigen::Ref<Eigen::MatrixXf> input) const {
   // instance of class std::normal_distribution with specific mean and stddev
   std::normal_distribution<float> distribution(0, sigma);
 
-  for (int i = 0; i < input.rows(); i++) {
-    for (int j = 0; j < input.cols(); j++) {
+  for (long i = 0; i < input.rows(); i++) {
+    for (long j = 0; j < input.cols(); j++) {
       input(i, j) = distribution(generator);
     }
   }
